refactor(react-web): extract useFocusTrap hook from PopupHost
Move the focus-trap effect (focus on activate, Tab/Shift+Tab cycle, restore on
close) into a shared useFocusTrap(ref, active) hook. PopupHost's PopupLayer now
uses it; behavior unchanged. Prepares a self-contained EngineErrorModal to reuse
the same trap instead of Modal's.
